# API/openapi-public.yaml
openapi: 3.0.3
info:
  title: Odyssey Finance API
  version: 1.1.0
  description: |
    오디세이 금융계획 웹 서비스의 외부 API (Spring). React 웹이 호출한다.

    ## 확정된 규약

    - **금액**: 원 단위 정수. 소수점 없음.
    - **비율**: 0~1 사이 소수. `0.15`가 15%다. 퍼센트 정수(`15`)를 보내면 안 된다.
    - **시각**: ISO 8601 + offset 필수. `2026-08-01T00:00:00+09:00` (O) / `2026-08-01` (X)
    - **날짜**: `date` 타입만 offset 없는 `YYYY-MM-DD` 사용 (생년월일, 목표일 등)
    - **월 표기**: `yearMonth`는 그 달 1일의 date. `2026-08-01` = 2026년 8월
    - **계획 생성**: 숫자·fan chart는 동기, LLM 설명만 Redis queue 후 별도 폴링

    ## 하이브리드 응답 방식

    `POST /goals/{goalId}/plan-versions`는 MC 시뮬레이션과 옵션 산출까지만 동기로
    처리하고 즉시 200을 반환한다. 정상 enqueue면 `explanation.status = PENDING`이고
    `explanation.text`는 null이다. Redis 발행이 실패하면 계획은 그대로 반환하고 설명만
    같은 응답에서 `FALLBACK`일 수 있다.

    LLM 설명은 백그라운드에서 생성되며, 클라이언트는
    `GET /plan-versions/{planVersionId}/explanation`을 2초 간격으로 폴링해
    `READY` 또는 `FALLBACK`이 될 때까지 기다린다. 설명이 실패해도 계획 자체는
    이미 사용 가능하다.

    ## 에러

    RFC 7807 `application/problem+json`. `code` 필드로 분기한다.
    409는 대부분 동시 요청 경합이므로, 클라이언트는 해당 리소스를 재조회하면 된다.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: 목표 목록
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, ACHIEVED, CANCELLED]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Goal' }
    post:
      operationId: createGoal
      tags: [goals]
      summary: 목표 생성
      description: |
        MVP는 사용자당 ACTIVE 목표를 하나로 제한한다
        (`uq_goal_active_per_user`). 이미 ACTIVE 목표가 있으면 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoalInput' }
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Goal' }
        '400':
          $ref: '#/components/responses/Problem400'
        '422':
          description: CUSTOM 금액이 확정 소비 하한보다 작음
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: 이미 ACTIVE 목표가 존재 (code=ACTIVE_GOAL_EXISTS)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /goals/{goalId}:
    parameters:
      - $ref: '#/components/parameters/GoalId'
    get:
      operationId: getGoal
      tags: [goals]
      summary: 목표 단건
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Goal' }
        '404':
          $ref: '#/components/responses/Problem404'
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: 목표 수정
      description: |
        `targetAmount` 또는 `targetDate`가 바뀌면 즉시 재계획 트리거
        (`GOAL_AMOUNT_CHANGED` / `GOAL_DATE_CHANGED`)가 발생한다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoalPatch' }
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Goal'
                  - type: object
                    properties:
                      triggeredReplanEventId:
                        type: integer
                        nullable: true
        '400':
          $ref: '#/components/responses/Problem400'
        '404':
          $ref: '#/components/responses/Problem404'

  # ─────────────────────────────────────────── scheduled expenses
  /scheduled-expenses:
    get:
      operationId: listScheduledExpenses
      tags: [scheduled-expenses]
      summary: 예정지출 목록
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PLANNED, COMPLETED, CANCELLED]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ScheduledExpense' }
    post:
      operationId: createScheduledExpense
      tags: [scheduled-expenses]
      summary: 예정지출 등록
      description: 등록 즉시 `SCHEDULED_EXPENSE_CHANGED` 재계획 트리거가 발생한다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduledExpenseInput' }
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ScheduledExpense'
                  - type: object
                    properties:
                      triggeredReplanEventId: { type: integer, nullable: true }
        '400':
          $ref: '#/components/responses/Problem400'
  /scheduled-expenses/{scheduledExpenseId}:
    parameters:
      - name: scheduledExpenseId
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateScheduledExpense
      tags: [scheduled-expenses]
      summary: 예정지출 수정 · 취소
      description: |
        `status=CANCELLED`로 바꾸면, 여기에 매칭돼 있던 거래는 일반 유동지출로
        되돌아가 부트스트랩 표본에 다시 포함된다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduledExpensePatch' }
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ScheduledExpense'
                  - type: object
                    properties:
                      triggeredReplanEventId: { type: integer, nullable: true }
        '404':
          $ref: '#/components/responses/Problem404'

  # ─────────────────────────────────────────── transactions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GoalId:
      name: goalId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Problem400:
      description: 요청 값 오류
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem404:
      description: 없거나 접근 권한 없음
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807
      required: [type, title, status, detail, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        code:
          type: string
          description: |
            클라이언트 분기용 코드.
            ACTIVE_GOAL_EXISTS · PROPOSED_PLAN_EXISTS · PLAN_INFEASIBLE ·
            PLAN_NOT_SELECTABLE · OPTION_ALREADY_SELECTED · DECISION_ALREADY_MADE ·
            ONBOARDING_INCOMPLETE · VALIDATION_FAILED
        errors:
          type: array
          description: 필드 단위 검증 오류
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    Goal:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        targetAmount: { type: integer, format: int64 }
        currentSavedAmount: { type: integer, format: int64 }
        targetDate: { type: string, format: date }
        status: { type: string, enum: [ACTIVE, ACHIEVED, CANCELLED] }
        remainingMonths:
          type: integer
          description: 오늘(KST) 기준 목표일까지 남은 개월 수
        spendingReplanSuppressedUntil:
          type: string
          format: date
          nullable: true
        createdAt: { type: string, format: date-time }
    GoalInput:
      type: object
      required: [name, targetAmount, targetDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        targetAmount: { type: integer, format: int64, minimum: 1 }
        currentSavedAmount: { type: integer, format: int64, minimum: 0, default: 0 }
        targetDate: { type: string, format: date }
    GoalPatch:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        targetAmount: { type: integer, format: int64, minimum: 1 }
        currentSavedAmount: { type: integer, format: int64, minimum: 0 }
        targetDate: { type: string, format: date }
        status: { type: string, enum: [ACTIVE, ACHIEVED, CANCELLED] }
    ScheduledExpense:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        amount: { type: integer, format: int64 }
        scheduledDate: { type: string, format: date }
        status: { type: string, enum: [PLANNED, COMPLETED, CANCELLED] }
        matchedTransactionCount:
          type: integer
          description: 이 예정지출에 매칭된 실제 거래 건수
        matchedAmount:
          type: integer
          format: int64
          description: 매칭된 거래 합계. 예상액과 다를 수 있다
    ScheduledExpenseInput:
      type: object
      required: [name, amount, scheduledDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        amount: { type: integer, format: int64, minimum: 1 }
        scheduledDate: { type: string, format: date }
    ScheduledExpensePatch:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        amount: { type: integer, format: int64, minimum: 1 }
        scheduledDate: { type: string, format: date }
        status: { type: string, enum: [PLANNED, COMPLETED, CANCELLED] }
...